fix circular dependency issue, move flipTrade into own module
