+++ server/modules/robot/robot.js
@@ -2,6 +2,7 @@ const botStatus = require('./botStatus');
 const toggleCoinbot = require('./toggleCoinbot');
 
 const robot = {
+    botStatus: botStatus,
     toggleCoinbot: toggleCoinbot
 }
